Bilibi : support more languages (#5248)
* Bilibi : support more languages

 They support more language than Hakuneko.  Currently we have 2 connectors : english and chinese. Urls are differents
Now they also have mangas in french, spanish, and indonesian. We just have to put a language code in one api request
The catch is : you get a different manga ID for a manga and a language. Therefore for gathering mangas list we have to performs loops with langage codes but there will be more manga displayed in list (if we make Bibili 'english' => multilingual')
Or we make different connectors with just a langage code different
But then, since url of website is the same, clipboard copy wont work as many website will have the exact same url