Fixes MangaCross plugin
The MangaCross plugin will fail if a url with an ending "/" is presented, since the ending "/" plus pop() will return a blank array. We can avoid this by filtering the array list to only show booleans before using pop(), thus always returning the manga id whether the URL is ending with a / or not. This code is probably not ideal so it's totally okay to decline and fix this with a better alternative.